+++ src/Framework/Gacela.php
@@ -7,6 +7,7 @@ namespace Gacela\Framework;
 use Gacela\Framework\Bootstrap\GacelaConfig;
 use Gacela\Framework\Bootstrap\SetupGacela;
 use Gacela\Framework\Bootstrap\SetupGacelaInterface;
+
 use function is_callable;
 
 final class Gacela
@@ -14,17 +15,39 @@ final class Gacela
     /**
      * Define the entry point of Gacela.
      *
-     * @param null|SetupGacelaInterface|callable(GacelaConfig):void $setup
+     * @param null|SetupGacelaInterface|callable(GacelaConfig):void $configFn SetupGacelaInterface is deprecated
      */
-    public static function bootstrap(string $appRootDir, $setup = null): void
+    public static function bootstrap(string $appRootDir, $configFn = null): void
     {
-        if (is_callable($setup)) {
-            $setup = SetupGacela::fromCallable($setup);
+        if ($configFn instanceof SetupGacelaInterface) {
+            trigger_deprecation(
+                'gacela-project/gacela',
+                '0.18',
+                '`SetupGacelaInterface` is deprecated. Use `callable(GacelaConfig)` instead.'
+            );
         }
+
+        $setup = self::normalizeSetup($configFn);
 
         Config::getInstance()
             ->setAppRootDir($appRootDir)
-            ->setSetup($setup ?? new SetupGacela())
+            ->setSetup($setup)
             ->init();
+    }
+
+    /**
+     * @param null|SetupGacelaInterface|callable(GacelaConfig):void $configFn
+     */
+    private static function normalizeSetup($configFn): SetupGacelaInterface
+    {
+        if ($configFn === null) {
+            return new SetupGacela();
+        }
+
+        if (is_callable($configFn)) {
+            return SetupGacela::fromCallable($configFn);
+        }
+
+        return $configFn;
     }
 }